Implemented blocked-cache
A key block will takes 64B, the size of a cache line. With a spin lock
and a position specifier of the ring block, 2 cache line (128B) will be
taken for each key block.
Inside a block, it is guaranteed that when inserting key, the key being
removed is the lease inserted key (not the least assessed key, as LRU
does). LRU inside a block can also be enabled via macro
NATIVE_ARRANGE_LRU.